Release 0.1.04
* Added functionality to associations service to make a duplicate
collection of references to application context data to use for
searching / management of associations in the details panel
* General code clean up & refactoring of *-item.drv.ts elements for
item selection toggling
* Bug fixes
* Delete association between two entities
* Adds hover functionality to Actionable objects and collections
* Implements beginning of association management per entity, which
includes a configurable list of associable entities for each associable
entity that controls which types of entities that are associable o.O
* Refactored expandable / details directives to allow for
cross-directive management (such as opening automatically when an
element is selected within the UI, and closing automatically when
deleting the element)
* Associations input / tagging functionality which includes
* basic keyboard navigation of possible associations list
* view associated elements for the selected entity, grouped by data
type
* create and remove associations directly within the UI
* bug fixes
* UI updates
* extra directives for additional functionality
* Added associations for leagues and conferences
* logo placeholders for league and conference
* league navigation / element details selection
* css tweaks
* added functionality for details selection & management for leagues
and conferences
* League module UI updates
* Fixed sidebar bug
* added details panel for browser and drill down
* Implements a conference-item directive to control rendering,
selection, and interaction of conference elements within the UI
* Conference model added
* Conference Collection added
* Added Conference API support
* Added Conference API constants
* Conference modal dialogs: initial implementation of Conference modal
dialogs for create, delete, and duplication error.
* Base conference UI started…
* added ability to open a scenario from the ‘new tab’ button in the
editor

+++ common/models/Actionable.ts
@@ -13,6 +13,7 @@ module Common.Models {
 		public disabled: boolean;
 		public clickable: boolean;
 		public hoverable: boolean;
+        public hovered: boolean;
 		public selected: boolean;
 		public selectable: boolean;
         public dragging: boolean;
@@ -31,6 +32,7 @@ module Common.Models {
             this.selected = false;
             this.clickable = true;
             this.hoverable = true;
+            this.hovered = false;
             this.dragging = false;
             this.draggable = true;
             this.selectable = true;
@@ -140,44 +142,62 @@ module Common.Models {
 
         public drop(): void {
             this.dragging = false;
-            this.graphics.drop();
+
+            if(this.hasGraphics())
+                this.graphics.drop();
         }
 
         public onhover(hoverIn: any, hoverOut: any, context: Common.Interfaces.IActionable): void {
-            this.graphics.onhover(hoverIn, hoverOut, context);
+            if (this.hasGraphics())
+                this.graphics.onhover(hoverIn, hoverOut, context);
         }
         public hoverIn(e: any) {
-            this.graphics.hoverIn(e);
+            this.hovered = true;
+
+            if(this.hasGraphics())
+                this.graphics.hoverIn(e);
         }
         public hoverOut(e: any) {
-            this.graphics.hoverOut(e);
+            this.hovered = false;
+
+            if(this.hasGraphics())
+                this.graphics.hoverOut(e);
         }
         public onclick(fn: any, context: Common.Interfaces.IActionable): void {
-            this.graphics.onclick(fn, context);
+            if (this.hasGraphics())
+                this.graphics.onclick(fn, context);
         }
         public click(e: any): void {
-            this.graphics.click(e);
+            if (this.hasGraphics())
+                this.graphics.click(e);
         }
         public oncontextmenu(fn: any, context: Common.Interfaces.IActionable): void {
-            this.graphics.oncontextmenu(fn, context);
+            if (this.hasGraphics())
+                this.graphics.oncontextmenu(fn, context);
         }
         public contextmenu(e: any): void {
-            this.graphics.contextmenu(e);
+            if (this.hasGraphics())
+                this.graphics.contextmenu(e);
         }
         public onmousedown(fn: any, context: Common.Interfaces.IActionable): void {
-            this.graphics.onmousedown(fn, context);
+            if (this.hasGraphics())
+                this.graphics.onmousedown(fn, context);
         }
         public onmouseup(fn: any, context: Common.Interfaces.IActionable): void {
-            this.graphics.onmouseup(fn, context);
+            if(this.hasGraphics())
+                this.graphics.onmouseup(fn, context);
         }
         public mousedown(e: any): void {
-            this.graphics.mousedown(e);
+            if (this.hasGraphics())
+                this.graphics.mousedown(e);
         }
         public onmousemove(fn: any, context: Common.Interfaces.IActionable): void {
-            this.graphics.onmousemove(fn, context);
+            if(this.hasGraphics())
+                this.graphics.onmousemove(fn, context);
         }
         public mousemove(e: any): void {
-            this.graphics.mousemove(e);
+            if (this.hasGraphics())
+                this.graphics.mousemove(e);
         }
         public ondrag(
             dragMove: Function,
@@ -185,7 +205,8 @@ module Common.Models {
             dragEnd: Function,
             context: Common.Interfaces.IActionable
         ): void {
-            this.graphics.ondrag(dragMove, dragStart, dragEnd, context);
+            if(this.hasGraphics())
+                this.graphics.ondrag(dragMove, dragStart, dragEnd, context);
         }
 
 	}
